[Product] Refactor product variant generator
