+++ src/Sylius/Bundle/ProductBundle/Resources/config/services.xml
@@ -13,8 +13,6 @@
 
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
-        <service id="sylius.set_builder.cartesian" class="Sylius\Component\Product\SetBuilder\CartesianSetBuilder" />
-
         <service id="sylius.form.listener.product_variant_generator" class="Sylius\Bundle\ProductBundle\Form\EventSubscriber\GenerateProductVariantsSubscriber">
             <argument type="service" id="sylius.generator.product_variant" />
         </service>
@@ -25,7 +23,6 @@
 
         <service id="sylius.generator.product_variant" class="Sylius\Component\Product\Generator\ProductVariantGenerator">
             <argument type="service" id="sylius.factory.product_variant" />
-            <argument type="service" id="sylius.set_builder.cartesian" />
         </service>
 
         <service id="sylius.form.type.product_option_value_choice" class="Sylius\Bundle\ProductBundle\Form\Type\ProductOptionValueChoiceType">
